requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.2
Django==3.2.23
django-admin-honeypot-updated-2021==1.2.0
django-honeypot==1.1.0
django-ipware==6.0.3
djangorestframework==3.14.0
pillow==10.2.0
pyasn1==0.5.1
pycparser==2.21
python-ipware==2.0.1
pytz==2023.4
rsa==4.9
sqlparse==0.4.4
tzdata==2023.4
